+++ mypyc/codegen/emitwrapper.py
